# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Food Delivery API",
    "version": "0.9.0",
    "description": "API for customer ordering, restaurant operations, shipper delivery, payments and administration."
  },
  "servers": [{ "url": "http://localhost:3000" }],
...
  "paths": {
...
    "/api/jobs/imports": { "post": { "tags": ["Jobs"], "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportRequest" } } } }, "responses": { "201": { "description": "Import result" } } }, "get": { "tags": ["Jobs"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Import jobs" } } } },
    "/api/jobs/exports": { "post": { "tags": ["Jobs"], "security": [{ "bearerAuth": [] }], "responses": { "201": { "description": "Export job" } } }, "get": { "tags": ["Jobs"], "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Export jobs" } } } },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
    "schemas": {
...
      "ImportRequest": { "type": "object", "required": ["jobType", "content"], "properties": { "jobType": { "type": "string", "enum": ["MENU_ITEMS", "RESTAURANTS"] }, "format": { "type": "string", "enum": ["CSV", "EXCEL"], "description": "EXCEL accepts SpreadsheetML .xls/.xml text in this offline build." }, "fileName": { "type": "string" }, "content": { "type": "string" } } }
    }
  }
}